HYD-4559 Integrate Req module into view server
The view server is currently using the requestStream module, but we
should instead be using the Req module now that we have extracted the
logic out of the realtime service.

To facilitate this change, we will create an api-request.js file, which
will take care of making the request through the req module. It will
take care of constructing the serverOptions object that is needed by the
req module. All occurrences of requestStream should instead be replaced
with apiRequest.

- Added api-request.js to handle making the api call using the req
  module
- Updated conf.js such that the requestUrl doesn't include the '/api/'
  in the url (since api-request.js takes care of this). Also updated
  one of the names from parsedApiHttpUrl to parsedServerHttpUrl.
- Updated index.js to use api-request instead of request-stream.
  This made it a bit cleaner.
- Replaced all occurrences of requestStream to apiRequest.
- Updated package.json to include the req module and removed
  request-stream from dependency list.

Signed-off-by: Joe Grund <joe.grund@intel.com>