HYD-5783 Update view-server to cache lnet_configuration data
HYD-5776 addressed the issue of adding lnetConfiguration to the store
and pulling data from the store in both the server and server details
pages. The next logical step is to cache the lnet_configuration data in
the view server.

In this patch the view server will be updated to cache lnet
configuration data so that it can be retrieve by the store quickly.

- Added lnet_configuration to the getCache call
- Added /lnet_configuration to the get-cache test

Signed-off-by: Will Johnson <william.c.johnson@intel.com>