Rely on supervisor-status daemon
We are currently determining supervisor status
by parsing an xmlrpc call.

In the interest of moving in a microservices type direction, we've moved the munging to it's own service that can be read via unix domain socket.

remove the code that connects via xmlrpc and just read from the service instead.